Moved quad_kers to quad_ker.py, changed imports
