+++ src/eko/evolution_operator/quad_ker.py
@@ -1,13 +1,20 @@
 r"""Integration kernels."""
 
+import enum
 import logging
 
 import numba as nb
 import numpy as np
 
-from .. import interpolation
+import ekore.anomalous_dimensions.polarized.space_like as ad_ps
+import ekore.anomalous_dimensions.unpolarized.space_like as ad_us
+import ekore.anomalous_dimensions.unpolarized.time_like as ad_ut
+import ekore.operator_matrix_elements.polarized.space_like as ome_ps
+import ekore.operator_matrix_elements.unpolarized.space_like as ome_us
+import ekore.operator_matrix_elements.unpolarized.time_like as ome_ut
+
+from .. import interpolation, mellin
 from .. import scale_variations as sv
-from ..io.types import InversionMethod
 from ..kernels import non_singlet as ns
 from ..kernels import non_singlet_qed as qed_ns
 from ..kernels import singlet as s
@@ -16,6 +23,7 @@ from ..kernels import valence_qed as qed_v
 from ..matchings import lepton_number
 from ..scale_variations import expanded as sv_expanded
 from ..scale_variations import exponentiated as sv_exponentiated
+from ..scale_variations.exponentiated import gamma_variation
 
 logger = logging.getLogger(__name__)
 
@@ -41,270 +49,6 @@ def select_singlet_element(ker, mode0, mode1):
     j = 0 if mode0 == 100 else 1
     k = 0 if mode1 == 100 else 1
     return ker[j, k]
-
-
-CB_SIGNATURE = nb.types.double(
-    nb.types.CPointer(nb.types.double),  # re_*_raw
-    nb.types.CPointer(nb.types.double),  # im_*_raw
-    nb.types.double,  # re_n
-    nb.types.double,  # im_n
-    nb.types.double,  # re_jac
-    nb.types.double,  # im_jac
-    nb.types.uintc,  # order_qcd
-    nb.types.uintc,  # order_qed
-    nb.types.bool_,  # is_singlet
-    nb.types.uintc,  # mode0
-    nb.types.uintc,  # mode1
-    nb.types.uintc,  # nf
-    nb.types.bool_,  # is_log
-    nb.types.double,  # logx
-    nb.types.CPointer(nb.types.double),  # areas_raw
-    nb.types.uintc,  # areas_x
-    nb.types.uintc,  # areas_y
-    nb.types.uintc,  # method_num
-    nb.types.double,  # as1
-    nb.types.double,  # as0
-    nb.types.uintc,  # ev_op_iterations
-    nb.types.uintc,  # ev_op_max_order_qcd
-    nb.types.uintc,  # sv_mode_num
-    nb.types.bool_,  # is_threshold
-    nb.types.double,  # Lsv
-    nb.types.CPointer(nb.types.double),  # as_list
-    nb.types.uintc,  # as_list_len
-    nb.types.double,  # mu2_from
-    nb.types.double,  # mu2_to
-    nb.types.CPointer(nb.types.double),  # a_half
-    nb.types.uintc,  # a_half_x
-    nb.types.uintc,  # a_half_y
-    nb.types.bool_,  # alphaem_running
-)
-
-
-@nb.cfunc(
-    CB_SIGNATURE,
-    cache=True,
-    nopython=True,
-)
-def cb_quad_ker_qcd(
-    re_gamma_raw,
-    im_gamma_raw,
-    re_n,
-    im_n,
-    re_jac,
-    im_jac,
-    order_qcd,
-    _order_qed,
-    is_singlet,
-    mode0,
-    mode1,
-    nf,
-    is_log,
-    logx,
-    areas_raw,
-    areas_x,
-    areas_y,
-    ev_method,
-    as1,
-    as0,
-    ev_op_iterations,
-    ev_op_max_order_qcd,
-    sv_mode,
-    is_threshold,
-    Lsv,
-    _as_list,
-    _as_list_len,
-    _mu2_from,
-    _mu2_to,
-    _a_half,
-    _a_half_x,
-    _a_half_y,
-    _alphaem_running,
-):
-    """C Callback inside integration kernel."""
-    # recover complex variables
-    n = re_n + im_n * 1j
-    jac = re_jac + im_jac * 1j
-    # combute basis functions
-    areas = nb.carray(areas_raw, (areas_x, areas_y))
-    pj = interpolation.evaluate_grid(n, is_log, logx, areas)
-    order = (order_qcd, 0)
-    ev_op_max_order = (ev_op_max_order_qcd, 0)
-    if is_singlet:
-        # reconstruct singlet matrices
-        re_gamma_singlet = nb.carray(re_gamma_raw, (order_qcd, 2, 2))
-        im_gamma_singlet = nb.carray(im_gamma_raw, (order_qcd, 2, 2))
-        gamma_singlet = re_gamma_singlet + im_gamma_singlet * 1j
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_singlet = sv_exponentiated.gamma_variation(
-                gamma_singlet, order, nf, Lsv
-            )
-        # construct eko
-        ker = s.dispatcher(
-            order,
-            ev_method,
-            gamma_singlet,
-            as1,
-            as0,
-            nf,
-            ev_op_iterations,
-            ev_op_max_order,
-        )
-        # scale var expanded is applied on the kernel
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = np.ascontiguousarray(
-                sv_expanded.singlet_variation(gamma_singlet, as1, order, nf, Lsv, dim=2)
-            ) @ np.ascontiguousarray(ker)
-        ker = select_singlet_element(ker, mode0, mode1)
-    else:
-        # construct non-singlet matrices
-        re_gamma_ns = nb.carray(re_gamma_raw, order_qcd)
-        im_gamma_ns = nb.carray(im_gamma_raw, order_qcd)
-        gamma_ns = re_gamma_ns + im_gamma_ns * 1j
-        if sv_mode == sv.Modes.exponentiated:
-            gamma_ns = sv_exponentiated.gamma_variation(gamma_ns, order, nf, Lsv)
-        # construct eko
-        ker = ns.dispatcher(
-            order,
-            ev_method,
-            gamma_ns,
-            as1,
-            as0,
-            nf,
-        )
-        if sv_mode == sv.Modes.expanded and not is_threshold:
-            ker = sv_expanded.non_singlet_variation(gamma_ns, as1, order, nf, Lsv) * ker
-    # recombine everything
-    res = ker * pj * jac
-    return np.real(res)
-
-
-@nb.njit(cache=True)
-def build_ome(A, matching_order, a_s, backward_method):
-    r"""Construct the matching expansion in :math:`a_s` with the appropriate
-    method.
-
-    Parameters
-    ----------
-    A : numpy.ndarray
-        list of |OME|
-    matching_order : tuple(int,int)
-        perturbation matching order
-    a_s : float
-        strong coupling, needed only for the exact inverse
-    backward_method : InversionMethod or None
-        empty or method for inverting the matching condition (exact or expanded)
-
-    Returns
-    -------
-    ome : numpy.ndarray
-        matching operator matrix
-    """
-    # to get the inverse one can use this FORM snippet
-    # Symbol a;
-    # NTensor c,d,e;
-    # Local x=-(a*c+a**2* d + a**3 * e);
-    # Local bi = 1+x+x**2+x**3;
-    # Print;
-    # .end
-    ome = np.eye(len(A[0]), dtype=np.complex128)
-    A = A[:, :, :]
-    A = np.ascontiguousarray(A)
-    if backward_method is InversionMethod.EXPANDED:
-        # expended inverse
-        if matching_order[0] >= 1:
-            ome -= a_s * A[0]
-        if matching_order[0] >= 2:
-            ome += a_s**2 * (-A[1] + A[0] @ A[0])
-        if matching_order[0] >= 3:
-            ome += a_s**3 * (-A[2] + A[0] @ A[1] + A[1] @ A[0] - A[0] @ A[0] @ A[0])
-    else:
-        # forward or exact inverse
-        if matching_order[0] >= 1:
-            ome += a_s * A[0]
-        if matching_order[0] >= 2:
-            ome += a_s**2 * A[1]
-        if matching_order[0] >= 3:
-            ome += a_s**3 * A[2]
-        # need inverse exact ?  so add the missing pieces
-        if backward_method is InversionMethod.EXACT:
-            ome = np.linalg.inv(ome)
-    return ome
-
-
-@nb.cfunc(
-    CB_SIGNATURE,
-    cache=True,
-    nopython=True,
-)
-def cb_quad_ker_ome(
-    re_ome_raw,
-    im_ome_raw,
-    re_n,
-    im_n,
-    re_jac,
-    im_jac,
-    order_qcd,
-    _order_qed,
-    is_singlet,
-    mode0,
-    mode1,
-    nf,
-    is_log,
-    logx,
-    areas_raw,
-    areas_x,
-    areas_y,
-    backward_method,
-    as1,
-    _as0,
-    _ev_op_iterations,
-    _ev_op_max_order_qcd,
-    sv_mode,
-    _is_threshold,
-    Lsv,
-    _as_list,
-    _as_list_len,
-    _mu2_from,
-    _mu2_to,
-    _a_half,
-    _a_half_x,
-    _a_half_y,
-    _alphaem_running,
-):
-    """C Callback inside integration kernel."""
-    # recover complex variables
-    n = re_n + im_n * 1j
-    jac = re_jac + im_jac * 1j
-    # compute basis functions
-    areas = nb.carray(areas_raw, (areas_x, areas_y))
-    pj = interpolation.evaluate_grid(n, is_log, logx, areas)
-    order = (order_qcd, 0)
-    if is_singlet:
-        indices = {21: 0, 100: 1, 90: 2}
-        # reconstruct singlet matrices
-        re_ome_singlet = nb.carray(re_ome_raw, (order_qcd, 3, 3))
-        im_ome_singlet = nb.carray(im_ome_raw, (order_qcd, 3, 3))
-        A = re_ome_singlet + im_ome_singlet * 1j
-    else:
-        indices = {200: 0, 91: 1}
-        # construct non-singlet matrices
-        re_ome_ns = nb.carray(re_ome_raw, (order_qcd, 2, 2))
-        im_ome_ns = nb.carray(im_ome_raw, (order_qcd, 2, 2))
-        A = re_ome_ns + im_ome_ns * 1j
-
-    # correct for scale variations
-    if sv_mode == sv.Modes.exponentiated:
-        A = sv.exponentiated.gamma_variation(A, order, nf, Lsv)
-
-    # build the expansion in alpha_s depending on the strategy
-    ker = build_ome(A, order, as1, backward_method)
-
-    # select the needed matrix element
-    ker = ker[indices[mode0], indices[mode1]]
-
-    # recombine everything
-    res = ker * pj * jac
-    return np.real(res)
 
 
 @nb.njit(cache=True)
@@ -365,110 +109,508 @@ def select_QEDvalence_element(ker, mode0, mode1):
     return ker[index1, index2]
 
 
-@nb.cfunc(
-    CB_SIGNATURE,
-    cache=True,
-    nopython=True,
-)
-def cb_quad_ker_qed(
-    re_gamma_raw,
-    im_gamma_raw,
-    re_n,
-    im_n,
-    re_jac,
-    im_jac,
-    order_qcd,
-    order_qed,
-    is_singlet,
+class MatchingMethods(enum.IntEnum):
+    """Enumerate matching methods."""
+
+    FORWARD = enum.auto()
+    BACKWARD_EXACT = enum.auto()
+    BACKWARD_EXPANDED = enum.auto()
+
+
+@nb.njit(cache=True)
+def build_ome(A, matching_order, a_s, backward_method):
+    r"""Construct the matching expansion in :math:`a_s` with the appropriate
+    method.
+
+    Parameters
+    ----------
+    A : numpy.ndarray
+        list of |OME|
+    matching_order : tuple(int,int)
+        perturbation matching order
+    a_s : float
+        strong coupling, needed only for the exact inverse
+    backward_method : MatchingMethods
+        empty or method for inverting the matching condition (exact or expanded)
+
+    Returns
+    -------
+    ome : numpy.ndarray
+        matching operator matrix
+    """
+    # to get the inverse one can use this FORM snippet
+    # Symbol a;
+    # NTensor c,d,e;
+    # Local x=-(a*c+a**2* d + a**3 * e);
+    # Local bi = 1+x+x**2+x**3;
+    # Print;
+    # .end
+    ome = np.eye(len(A[0]), dtype=np.complex128)
+    A = A[:, :, :]
+    A = np.ascontiguousarray(A)
+    if backward_method is MatchingMethods.BACKWARD_EXPANDED:
+        # expended inverse
+        if matching_order[0] >= 1:
+            ome -= a_s * A[0]
+        if matching_order[0] >= 2:
+            ome += a_s**2 * (-A[1] + A[0] @ A[0])
+        if matching_order[0] >= 3:
+            ome += a_s**3 * (-A[2] + A[0] @ A[1] + A[1] @ A[0] - A[0] @ A[0] @ A[0])
+    else:
+        # forward or exact inverse
+        if matching_order[0] >= 1:
+            ome += a_s * A[0]
+        if matching_order[0] >= 2:
+            ome += a_s**2 * A[1]
+        if matching_order[0] >= 3:
+            ome += a_s**3 * A[2]
+        # need inverse exact ?  so add the missing pieces
+        if backward_method is MatchingMethods.BACKWARD_EXACT:
+            ome = np.linalg.inv(ome)
+    return ome
+
+
+spec = [
+    ("is_singlet", nb.boolean),
+    ("is_QEDsinglet", nb.boolean),
+    ("is_QEDvalence", nb.boolean),
+    ("is_log", nb.boolean),
+    ("logx", nb.float64),
+    ("u", nb.float64),
+]
+
+
+@nb.experimental.jitclass(spec)
+class QuadKerBase:
+    """Manage the common part of Mellin inversion integral.
+
+    Parameters
+    ----------
+    u : float
+        quad argument
+    is_log : boolean
+        is a logarithmic interpolation
+    logx : float
+        Mellin inversion point
+    mode0 : str
+        first sector element
+    """
+
+    def __init__(self, u, is_log, logx, mode0):
+        self.is_singlet = mode0 in [100, 21, 90]
+        self.is_QEDsinglet = mode0 in [21, 22, 100, 101, 90]
+        self.is_QEDvalence = mode0 in [10200, 10204]
+        self.is_log = is_log
+        self.u = u
+        self.logx = logx
+
+    @property
+    def path(self):
+        """Return the associated instance of :class:`eko.mellin.Path`."""
+        if self.is_singlet or self.is_QEDsinglet:
+            return mellin.Path(self.u, self.logx, True)
+        else:
+            return mellin.Path(self.u, self.logx, False)
+
+    @property
+    def n(self):
+        """Returs the Mellin moment :math:`N`."""
+        return self.path.n
+
+    def integrand(
+        self,
+        areas,
+    ):
+        """Get transformation to Mellin space integral.
+
+        Parameters
+        ----------
+        areas : tuple
+            basis function configuration
+
+        Returns
+        -------
+        complex
+            common mellin inversion integrand
+        """
+        if self.logx == 0.0:
+            return 0.0
+        pj = interpolation.evaluate_grid(self.path.n, self.is_log, self.logx, areas)
+        if pj == 0.0:
+            return 0.0
+        return self.path.prefactor * pj * self.path.jac
+
+
+@nb.njit(cache=True)
+def quad_ker_ad(
+    u,
+    order,
     mode0,
     mode1,
-    nf,
+    method,
     is_log,
     logx,
-    areas_raw,
-    areas_x,
-    areas_y,
-    ev_method,
-    _as1,
-    _as0,
-    ev_op_iterations,
-    ev_op_max_order_qcd,
-    sv_mode,
-    is_threshold,
-    Lsv,
-    as_list_raw,
-    as_list_len,
+    areas,
+    as_list,
     mu2_from,
     mu2_to,
-    a_half_raw,
-    a_half_x,
-    a_half_y,
+    a_half,
     alphaem_running,
+    nf,
+    L,
+    ev_op_iterations,
+    ev_op_max_order,
+    sv_mode,
+    is_threshold,
+    n3lo_ad_variation,
+    is_polarized,
+    is_time_like,
+    use_fhmruvv,
 ):
-    """C Callback inside integration kernel."""
-    # recover complex variables
-    n = re_n + im_n * 1j
-    jac = re_jac + im_jac * 1j
-    # compute basis functions
-    areas = nb.carray(areas_raw, (areas_x, areas_y))
-    pj = interpolation.evaluate_grid(n, is_log, logx, areas)
-    order = (order_qcd, order_qed)
-    ev_op_max_order = (ev_op_max_order_qcd, order_qed)
-    is_valence = (mode0 == 10200) or (mode0 == 10204)
+    """Raw evolution kernel inside quad.
 
-    as_list = nb.carray(as_list_raw, as_list_len)
-    a_half = nb.carray(a_half_raw, (a_half_x, a_half_y))
+    Parameters
+    ----------
+    u : float
+        quad argument
+    order : int
+        perturbation order
+    mode0: int
+        pid for first sector element
+    mode1 : int
+        pid for second sector element
+    method : str
+        method
+    is_log : boolean
+        is a logarithmic interpolation
+    logx : float
+        Mellin inversion point
+    areas : tuple
+        basis function configuration
+    as1 : float
+        target coupling value
+    as0 : float
+        initial coupling value
+    mu2_from : float
+        initial value of mu2
+    mu2_from : float
+        final value of mu2
+    aem_list : list
+        list of electromagnetic coupling values
+    alphaem_running : bool
+        whether alphaem is running or not
+    nf : int
+        number of active flavors
+    L : float
+        logarithm of the squared ratio of factorization and renormalization scale
+    ev_op_iterations : int
+        number of evolution steps
+    ev_op_max_order : int
+        perturbative expansion order of U
+    sv_mode: int, `enum.IntEnum`
+        scale variation mode, see `eko.scale_variations.Modes`
+    is_threshold : boolean
+        is this an intermediate threshold operator?
+    n3lo_ad_variation : tuple
+        |N3LO| anomalous dimension variation ``(gg, gq, qg, qq, nsp, nsm, nsv)``
+    is_polarized : boolean
+        is polarized evolution ?
+    is_time_like : boolean
+        is time-like evolution ?
+    use_fhmruvv : bool
+        if True use the |FHMRUVV| |N3LO| anomalous dimension
 
-    if is_singlet:
-        # reconstruct singlet matrices
-        re_gamma_singlet = nb.carray(re_gamma_raw, (order_qcd + 1, order_qed + 1, 4, 4))
-        im_gamma_singlet = nb.carray(im_gamma_raw, (order_qcd + 1, order_qed + 1, 4, 4))
-        gamma_singlet = re_gamma_singlet + im_gamma_singlet * 1j
+    Returns
+    -------
+    float
+        evaluated integration kernel
+    """
+    ker_base = QuadKerBase(u, is_log, logx, mode0)
+    integrand = ker_base.integrand(areas)
+    if integrand == 0.0:
+        return 0.0
+    if order[1] == 0:
+        ker = quad_ker_qcd(
+            ker_base,
+            order,
+            mode0,
+            mode1,
+            method,
+            as_list[-1],
+            as_list[0],
+            nf,
+            L,
+            ev_op_iterations,
+            ev_op_max_order,
+            sv_mode,
+            is_threshold,
+            is_polarized,
+            is_time_like,
+            n3lo_ad_variation,
+            use_fhmruvv,
+        )
+    else:
+        ker = quad_ker_qed(
+            ker_base,
+            order,
+            mode0,
+            mode1,
+            method,
+            as_list,
+            mu2_from,
+            mu2_to,
+            a_half,
+            alphaem_running,
+            nf,
+            L,
+            ev_op_iterations,
+            ev_op_max_order,
+            sv_mode,
+            is_threshold,
+            n3lo_ad_variation,
+            use_fhmruvv,
+        )
 
+    # recombine everything
+    return np.real(ker * integrand)
+
+
+@nb.njit(cache=True)
+def quad_ker_qcd(
+    ker_base,
+    order,
+    mode0,
+    mode1,
+    method,
+    as1,
+    as0,
+    nf,
+    L,
+    ev_op_iterations,
+    ev_op_max_order,
+    sv_mode,
+    is_threshold,
+    is_polarized,
+    is_time_like,
+    n3lo_ad_variation,
+    use_fhmruvv,
+):
+    """Raw evolution kernel inside quad.
+
+    Parameters
+    ----------
+    quad_ker : float
+        quad argument
+    order : int
+        perturbation order
+    mode0: int
+        pid for first sector element
+    mode1 : int
+        pid for second sector element
+    method : str
+        method
+    as1 : float
+        target coupling value
+    as0 : float
+        initial coupling value
+    nf : int
+        number of active flavors
+    L : float
+        logarithm of the squared ratio of factorization and renormalization scale
+    ev_op_iterations : int
+        number of evolution steps
+    ev_op_max_order : int
+        perturbative expansion order of U
+    sv_mode: int, `enum.IntEnum`
+        scale variation mode, see `eko.scale_variations.Modes`
+    is_threshold : boolean
+        is this an itermediate threshold operator?
+    n3lo_ad_variation : tuple
+        |N3LO| anomalous dimension variation ``(gg, gq, qg, qq, nsp, nsm, nsv)``
+    use_fhmruvv : bool
+        if True use the |FHMRUVV| |N3LO| anomalous dimensions
+
+    Returns
+    -------
+    float
+        evaluated integration kernel
+    """
+    # compute the actual evolution kernel for pure QCD
+    if ker_base.is_singlet:
+        if is_polarized:
+            if is_time_like:
+                raise NotImplementedError("Polarized, time-like is not implemented")
+            else:
+                gamma_singlet = ad_ps.gamma_singlet(order, ker_base.n, nf)
+        else:
+            if is_time_like:
+                gamma_singlet = ad_ut.gamma_singlet(order, ker_base.n, nf)
+            else:
+                gamma_singlet = ad_us.gamma_singlet(
+                    order, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
+                )
         # scale var exponentiated is directly applied on gamma
         if sv_mode == sv.Modes.exponentiated:
-            gamma_singlet = sv.exponentiated.gamma_variation_qed(
-                gamma_singlet, order, nf, lepton_number(mu2_to), Lsv, alphaem_running
+            gamma_singlet = sv_exponentiated.gamma_variation(
+                gamma_singlet, order, nf, L
             )
+        ker = s.dispatcher(
+            order,
+            method,
+            gamma_singlet,
+            as1,
+            as0,
+            nf,
+            ev_op_iterations,
+            ev_op_max_order,
+        )
+        # scale var expanded is applied on the kernel
+        if sv_mode == sv.Modes.expanded and not is_threshold:
+            ker = np.ascontiguousarray(
+                sv_expanded.singlet_variation(gamma_singlet, as1, order, nf, L, dim=2)
+            ) @ np.ascontiguousarray(ker)
+        ker = select_singlet_element(ker, mode0, mode1)
+    else:
+        if is_polarized:
+            if is_time_like:
+                raise NotImplementedError("Polarized, time-like is not implemented")
+            else:
+                gamma_ns = ad_ps.gamma_ns(order, mode0, ker_base.n, nf)
+        else:
+            if is_time_like:
+                gamma_ns = ad_ut.gamma_ns(order, mode0, ker_base.n, nf)
+            else:
+                gamma_ns = ad_us.gamma_ns(
+                    order, mode0, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
+                )
+        if sv_mode == sv.Modes.exponentiated:
+            gamma_ns = sv_exponentiated.gamma_variation(gamma_ns, order, nf, L)
+        ker = ns.dispatcher(
+            order,
+            method,
+            gamma_ns,
+            as1,
+            as0,
+            nf,
+        )
+        if sv_mode == sv.Modes.expanded and not is_threshold:
+            ker = sv_expanded.non_singlet_variation(gamma_ns, as1, order, nf, L) * ker
+    return ker
 
+
+@nb.njit(cache=True)
+def quad_ker_qed(
+    ker_base,
+    order,
+    mode0,
+    mode1,
+    method,
+    as_list,
+    mu2_from,
+    mu2_to,
+    a_half,
+    alphaem_running,
+    nf,
+    L,
+    ev_op_iterations,
+    ev_op_max_order,
+    sv_mode,
+    is_threshold,
+    n3lo_ad_variation,
+    use_fhmruvv,
+):
+    """Raw evolution kernel inside quad.
+
+    Parameters
+    ----------
+    ker_base : QuadKerBase
+        quad argument
+    order : int
+        perturbation order
+    mode0: int
+        pid for first sector element
+    mode1 : int
+        pid for second sector element
+    method : str
+        method
+    as1 : float
+        target coupling value
+    as0 : float
+        initial coupling value
+    mu2_from : float
+        initial value of mu2
+    mu2_from : float
+        final value of mu2
+    aem_list : list
+        list of electromagnetic coupling values
+    alphaem_running : bool
+        whether alphaem is running or not
+    nf : int
+        number of active flavors
+    L : float
+        logarithm of the squared ratio of factorization and renormalization scale
+    ev_op_iterations : int
+        number of evolution steps
+    ev_op_max_order : int
+        perturbative expansion order of U
+    sv_mode: int, `enum.IntEnum`
+        scale variation mode, see `eko.scale_variations.Modes`
+    is_threshold : boolean
+        is this an itermediate threshold operator?
+    n3lo_ad_variation : tuple
+        |N3LO| anomalous dimension variation ``(gg, gq, qg, qq, nsp, nsm, nsv)``
+    use_fhmruvv : bool
+        if True use the |FHMRUVV| |N3LO| anomalous dimensions
+
+    Returns
+    -------
+    float
+        evaluated integration kernel
+    """
+    # compute the actual evolution kernel for QEDxQCD
+    if ker_base.is_QEDsinglet:
+        gamma_s = ad_us.gamma_singlet_qed(
+            order, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
+        )
+        # scale var exponentiated is directly applied on gamma
+        if sv_mode == sv.Modes.exponentiated:
+            gamma_s = sv_exponentiated.gamma_variation_qed(
+                gamma_s, order, nf, lepton_number(mu2_to), L, alphaem_running
+            )
         ker = qed_s.dispatcher(
             order,
-            ev_method,
-            gamma_singlet,
+            method,
+            gamma_s,
             as_list,
             a_half,
             nf,
             ev_op_iterations,
             ev_op_max_order,
         )
+        # scale var expanded is applied on the kernel
+        # TODO : in this way a_half[-1][1] is the aem value computed in
+        # the middle point of the last step. Instead we want aem computed in mu2_final.
+        # However the distance between the two is very small and affects only the running aem
         if sv_mode == sv.Modes.expanded and not is_threshold:
             ker = np.ascontiguousarray(
-                sv.expanded.singlet_variation_qed(
-                    gamma_singlet,
-                    as_list[-1],
-                    a_half[-1][1],
-                    alphaem_running,
-                    order,
-                    nf,
-                    Lsv,
+                sv_expanded.singlet_variation_qed(
+                    gamma_s, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
                 )
             ) @ np.ascontiguousarray(ker)
         ker = select_QEDsinglet_element(ker, mode0, mode1)
-
-    elif is_valence:
-        # reconstruct valence matrices
-        re_gamma_valence = nb.carray(re_gamma_raw, (order_qcd + 1, order_qed + 1, 2, 2))
-        im_gamma_valence = nb.carray(im_gamma_raw, (order_qcd + 1, order_qed + 1, 2, 2))
-        gamma_valence = re_gamma_valence + im_gamma_valence * 1j
-
+    elif ker_base.is_QEDvalence:
+        gamma_v = ad_us.gamma_valence_qed(
+            order, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
+        )
+        # scale var exponentiated is directly applied on gamma
         if sv_mode == sv.Modes.exponentiated:
-            gamma_valence = sv.exponentiated.gamma_variation_qed(
-                gamma_valence, order, nf, lepton_number(mu2_to), Lsv, alphaem_running
+            gamma_v = sv_exponentiated.gamma_variation_qed(
+                gamma_v, order, nf, lepton_number(mu2_to), L, alphaem_running
             )
         ker = qed_v.dispatcher(
             order,
-            ev_method,
-            gamma_valence,
+            method,
+            gamma_v,
             as_list,
             a_half,
             nf,
@@ -478,31 +620,23 @@ def cb_quad_ker_qed(
         # scale var expanded is applied on the kernel
         if sv_mode == sv.Modes.expanded and not is_threshold:
             ker = np.ascontiguousarray(
-                sv.expanded.valence_variation_qed(
-                    gamma_valence,
-                    as_list[-1],
-                    a_half[-1][1],
-                    alphaem_running,
-                    order,
-                    nf,
-                    Lsv,
+                sv_expanded.valence_variation_qed(
+                    gamma_v, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
                 )
             ) @ np.ascontiguousarray(ker)
         ker = select_QEDvalence_element(ker, mode0, mode1)
-
     else:
-        # construct non-singlet matrices
-        re_gamma_ns = nb.carray(re_gamma_raw, (order_qcd + 1, order_qed + 1))
-        im_gamma_ns = nb.carray(im_gamma_raw, (order_qcd + 1, order_qed + 1))
-        gamma_ns = re_gamma_ns + im_gamma_ns * 1j
+        gamma_ns = ad_us.gamma_ns_qed(
+            order, mode0, ker_base.n, nf, n3lo_ad_variation, use_fhmruvv
+        )
+        # scale var exponentiated is directly applied on gamma
         if sv_mode == sv.Modes.exponentiated:
             gamma_ns = sv_exponentiated.gamma_variation_qed(
-                gamma_ns, order, nf, lepton_number(mu2_to), Lsv, alphaem_running
+                gamma_ns, order, nf, lepton_number(mu2_to), L, alphaem_running
             )
-        # construct eko
         ker = qed_ns.dispatcher(
             order,
-            ev_method,
+            method,
             gamma_ns,
             as_list,
             a_half[:, 1],
@@ -515,16 +649,107 @@ def cb_quad_ker_qed(
         if sv_mode == sv.Modes.expanded and not is_threshold:
             ker = (
                 sv_expanded.non_singlet_variation_qed(
-                    gamma_ns,
-                    as_list[-1],
-                    a_half[-1][1],
-                    alphaem_running,
-                    order,
-                    nf,
-                    Lsv,
+                    gamma_ns, as_list[-1], a_half[-1][1], alphaem_running, order, nf, L
                 )
                 * ker
             )
+    return ker
+
+
+@nb.njit(cache=True)
+def quad_ker_ome(
+    u,
+    order,
+    mode0,
+    mode1,
+    is_log,
+    logx,
+    areas,
+    a_s,
+    nf,
+    L,
+    sv_mode,
+    Lsv,
+    backward_method,
+    is_msbar,
+    is_polarized,
+    is_time_like,
+):
+    r"""Raw kernel inside quad.
+
+    Parameters
+    ----------
+    u : float
+        quad argument
+    order : tuple(int,int)
+        perturbation matching order
+    mode0 : int
+        pid for first element in the singlet sector
+    mode1 : int
+        pid for second element in the singlet sector
+    is_log : boolean
+        logarithmic interpolation
+    logx : float
+        Mellin inversion point
+    areas : tuple
+        basis function configuration
+    a_s : float
+        strong coupling, needed only for the exact inverse
+    nf: int
+        number of active flavor below threshold
+    L : float
+        :math:``\ln(\mu_F^2 / m_h^2)``
+    backward_method : InversionMethod or None
+        empty or method for inverting the matching condition (exact or expanded)
+    is_msbar: bool
+        add the |MSbar| contribution
+    is_polarized : boolean
+        is polarized evolution ?
+    is_time_like : boolean
+        is time-like evolution ?
+
+    Returns
+    -------
+    ker : float
+        evaluated integration kernel
+    """
+    ker_base = QuadKerBase(u, is_log, logx, mode0)
+    integrand = ker_base.integrand(areas)
+    if integrand == 0.0:
+        return 0.0
+    # compute the ome
+    if ker_base.is_singlet or ker_base.is_QEDsinglet:
+        indices = {21: 0, 100: 1, 90: 2}
+        if is_polarized:
+            if is_time_like:
+                raise NotImplementedError("Polarized, time-like is not implemented")
+            A = ome_ps.A_singlet(order, ker_base.n, nf, L)
+        else:
+            if is_time_like:
+                A = ome_ut.A_singlet(order, ker_base.n, L)
+            else:
+                A = ome_us.A_singlet(order, ker_base.n, nf, L, is_msbar)
+    else:
+        indices = {200: 0, 91: 1}
+        if is_polarized:
+            if is_time_like:
+                raise NotImplementedError("Polarized, time-like is not implemented")
+            A = ome_ps.A_non_singlet(order, ker_base.n, L)
+        else:
+            if is_time_like:
+                A = ome_ut.A_non_singlet(order, ker_base.n, L)
+            else:
+                A = ome_us.A_non_singlet(order, ker_base.n, nf, L)
+
+    # correct for scale variations
+    if sv_mode == sv.Modes.exponentiated:
+        A = gamma_variation(A, order, nf, Lsv)
+
+    # build the expansion in alpha_s depending on the strategy
+    ker = build_ome(A, order, a_s, backward_method)
+
+    # select the needed matrix element
+    ker = ker[indices[mode0], indices[mode1]]
+
     # recombine everything
-    res = ker * pj * jac
-    return np.real(res)
+    return np.real(ker * integrand)
